backend/requirements.txt
pillow~=10.3.0
psycopg2~=2.9.9

Flask~=3.0.3
requests~=2.32.3
tortoise-orm
asyncpg
fastapi[all]